Move get_objective utils to init to avoid circular imports
